Move SyketilfelleKafkaConsumer and friends to package kafka.consumers.syketilfelle
