+++ src/main/kotlin/no/nav/syfo/kafka/consumers/syketilfelle/domain/KSyketilfellebit.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.kafka.consumers.oppfolgingstilfelle.domain
+package no.nav.syfo.kafka.consumers.syketilfelle.domain
 
 import java.time.LocalDate
 import java.time.OffsetDateTime
